Extract shared GitHub I/O plumbing into pr_github.py (#600 §5)
Second member of the shared lib, on top of gh_cli (#691): the GraphQL executor,
the PR fetch, and the authenticated-actor lookup (_graphql, _fetch_pr,
_viewer_login) move into pr_github.py, built on gh_cli.run. These are the I/O
layer both the review-state projector and the thread-witness looker need
(§5 "_fetch_pr, thread walking … imported by both, duplicated by neither"), and
the prerequisite for lifting the looker subsystem cleanly next.

The engine imports them back, so its surface and the ~40 tests that mock at the
engine boundary are unchanged — except one: test_arm_auto_merge_degrades patched
_run to cover the _graphql call inside _auto_merge_state, which now runs through
gh_cli.run; it mocks _auto_merge_state directly instead (matching the sibling arm
tests) to isolate its actual subject, the merge-call authz path. Functions moved
verbatim. Engine suite 100/100, ruff clean.
